2002-05-28  Roland McGrath  <roland@frob.com>
	* dir.c (diskfs_get_directs): u_int -> size_t
	* tmpfs.c (diskfs_append_args): unsigned -> size_t

+++ tmpfs/tmpfs.c
@@ -231,7 +231,7 @@ parse_opt (int key, char *arg, struct argp_state *state)
 
 /* Override the standard diskfs routine so we can add our own output.  */
 error_t
-diskfs_append_args (char **argz, unsigned *argz_len)
+diskfs_append_args (char **argz, size_t *argz_len)
 {
   error_t err;
